requirements.txt
selenium>=4.18.0
webdriver-manager>=4.0.1
beautifulsoup4>=4.12.0
requests>=2.31.0
flask>=3.0.0
flask-cors>=4.0.0
openai>=1.14.0
python-dotenv>=1.0.0
lxml>=5.1.0
